+++ workalendar/europe/russia.py
@@ -26,6 +26,12 @@ class Russia(OrthodoxCalendar):
     covid19_2020_end = date(2020, 4, 30)
 
     def get_fixed_holidays(self, year):
+
+        if year >= 1992:
+            self.labour_day_label = "The Day of Spring and Labour"
+        else:
+            self.labour_day_label = "International Workers' Day"
+
         days = super().get_fixed_holidays(year)
 
         if year >= 2005:
@@ -61,6 +67,12 @@ class Russia(OrthodoxCalendar):
 
             # Constitution Vote Public Holiday
             days.append((date(year, 7, 1), "Constitution Vote Public Holiday"))
+        elif year == 2021:
+            days.extend([
+                (date(year, 2, 22), "Day Before Defendence of the Fatherland"),
+                (date(year, 11, 5), 'Day After Day of Unity'),
+                (date(year, 12, 31), "New Year's Eve"),
+            ])
         return days
 
     def get_calendar_holidays(self, year):